@@ -5,7 +5,7 @@ import { CacheService } from '../../imports';
 import { ChatService } from '../../services/chat.service';
 import { ChatRequestService } from '../../services/chat-request.service';
 
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-room',
