@@ -5,10 +5,10 @@ import { HelpersService } from './helpers.service';
 import { RoutingListService } from './routing-list.service';
 import { environment } from '../../../environments/environment';
 
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/switchMap';
+
+
+
+
 
 @Injectable()
 export class MainRequestService {
